Fixed 'different engine' save error
I changed the saving code to use GAMENAME instead of GAMESIG; I forgot to change the loading equivalent, resulting in an error when it compared the ''Engine'' text in the PNG save file.

To be precise, it saved ''The engine powering World of Kirbycraft, ZDoom,'' instead of ''World of Kirbycraft'', which the loading code checked for.

This was changed to be consistent.