Modified two things.
You didn't expect another commit, did you? I managed to fix the problem Git created and a new commit is here!

It modifies the default automap overlay colors to fit World of Kirbycraft (the same ones seen in my videos) and removes the water-running code.

That code was moved to ACS and will be in the next release.

Also, MinSizeRel generates a smaller EXE so that'll be used from now on.