Disabled making PK3's.
The reason for this is because for some reason the game couldn't read it correctly and would assume "actors/constants.txt" didn't exist, even if I manually made the PK3 with 7-Zip!

For whatever reason, GZDoom's official PK3 works if modified by SLADE 3. Not a fan of having to do this, but I've never seen the generated file be rejected by GZDoom befofe...what did the developers do differently than my version of the PK3?