Added latest GZDoom commits to https://github.com/coelckers/gzdoom/commit/dd8a114bb821ccc5e3a15e34151de71a3d907ffd.
Yes, I should've just made this a fork but I didn't do that YEARS ago so we're stuck with this awkward method of keeping up to date! I don't want to ose the history, terrible as it was, with this game.

Includes commits:
https://github.com/coelckers/gzdoom/commit/b6b78176d26f1a11f1e17b5f88a169c639f4138d
https://github.com/coelckers/gzdoom/commit/59ee89b622040f504ecd3aa59a63c6dc1a2c7b00
https://github.com/coelckers/gzdoom/commit/afdeed305bccc786fd019c191896426c8c1c87f6
https://github.com/coelckers/gzdoom/commit/dd8a114bb821ccc5e3a15e34151de71a3d907ffd

Hope none of dpjudas' edits break the game in the future...always worried about updating this game's engine...