Reapplied MMA commit https://github.com/Xane123/MaryMagicalAdventure/commit/76e1ee10851531ec4f08c6bd2017344fdd9a3070#diff-9462481aee4bd6d32c9f31f3fe24ea39
...minus stupid .IPCH files, of course, but I noticed the turrets had strange behavior, so I re-added this, along with the better-written error message.

Those developers need to update these errors to not sound as technical and "only-for-the-developers"; New errors have punctuation and look good. Older ones are crappy, forgot punctuation, and like to remind you what part of the code that called it. Those developers could learn from my style in p_enemy.cpp here.