Removed mention of "Doom" in cheats and fixed music cheat:musical_note:
Yes, you can request any song in the game as long as it exists! Type PLAY then a two-digit number to switch the current song. It's the same thing as "IDMUS" in Doom, but it instead runs the ACS script that handles changing songs!

Also, note the check for '0' in Cht_Music; I added that because I don't know if the game would know that "01" = "1" and so on until 10, so I just make it use the second number you pushed.