Changed internal name and "description" & slightly altered debug dialog size.
Yes, this fixes the weirdly-long name you may see in problem reports to Microsoft on Windows. This doesn't affect Mac OS or Linux as they get the title through other methods.

Note to other developers: Keep in mind "description" in your .rc files is really the title of the program that Windows uses to name it. Confusing why they didn't call "InternalName" "Title" and make "description" really be the "description" of the program...