:octocat: More ported changes!
Soon MMA will be just as it was and even better due to Delta Touch compatibilty! No, really, try downloading the game data then put it in your OpenTouch/Delta/mod folder. Select the three files in Delta Touch as mods then run the game and you'll be able to play it!