[ACS/DECORATE]Removed TOD system plus misc. fixes.
Yes, the time-of-day system caused too many aesthetic problems to the game so I had to remove it. Some levels will still change mid-way through them, similar to the system.

I may reuse code from this in the future if I want to make the Small Hut/Minigame Building inherit the time of day/weather from the level the player was just in.