That worked, so more splitting is in effect now!
Oddly, this is a lot like older versions, back in World of Kirbycraft. In those times, I used two files, "wok_x.pk3" and "wok_extra_lag.pk3", the latter named because of how slow it was to save with all of the data inside the file. I suppose SLADE 3 improved that, but now we're back to that, with not two but THREE files!