Really improved "stat" and "vid_fps" debug info's appearance!
With the stat display, there's a background behind all text, ensuring it's always readable. Also, different stats use different colors.

For the FPS counter, a similar gradient background is used, it's on the left side now, and it changes color based on FPS now. Set vid_fps to 2 to see the "ms" counter, which is otherwise not shown.