Reverted new :video_game: gamepad button names (as they weren't being saved).
Also, the game now doesn't pause if you bring down the console now. This may remind you of old World of Kirbycraft builds, but those did this with the menu. I wanted to do this with the console but didn't know where to look until now.

The menu still pauses as it's like the "pause menu". The console isn't, so eh.