:m: - Updated macOS icon and .plist.
I don't know what the .plist file really does as I'm a Windows user, but I've removed its file associations (as Magical Mary isn't meant to be used like GZDoom, where it's meant to load mods and stuff.

The ICNS, though, has been a while coming; I've wanted to replace it with the same icon as Windows, but I couldn't find a free converter. Due to this, I used Sonic Mania's macOS icon in earlier game versions. A Doom space marine or "GZ" in a circle wouldn't represent the speedy gameplay of this game as well as Sonic, after all.