Updated workflow to download additional Windows files.
Whenever GitHub Actions would build the Windows version of Magical Mary, it wouldn't include the DLL files that it needs to run. However, those are conveniently hosted in the cloud, so I can download them then extract them into the packaging directory.

This commit adds a command to download the ZIP containing the DLL files (and Linux's library), extract that, then it only compresses the OS-specific sub-directory. What a nice update!